Refactoring from Java to Kotlin + added packages
All the code that was previously in Java is now in Kotlin. Classes whos primary purpose is handling MQTT have been placed in a package named mqtt.
A package named activities has been created in which our screens can be placed.
A class called GameActivity.kt has been created which handles the game screen, i.e. displaying the buttons and camera and the users touch inputs.